# application/services/custom_services/ticket/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: meanappdemo57 open api specification
   description: ticket API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:8017
     description: ticket
tags:
- name: ticket
paths:
   "/ticket/{id}":
      delete:
         tags:
         -  ticket
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  ticket
         description: GpGetNounById
         parameters:
         -  in: "path"
            description: GpGetNounById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/ticket/get/search":
      get:
         tags:
         -  ticket
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/ticket/get/update":
      put:
         tags:
         -  ticket
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/ticket"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/ticket":
      put:
         tags:
         -  ticket
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/ticket"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  ticket
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  ticket
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/ticket"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/ticket/userid/created_by":
      get:
         tags:
         -  ticket
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      ticket:
         type: object
         properties:
            name:
               type: "string"
               description: Description
            description:
               type: "string"
               description: Description
            callername:
               type: "string"
               description: Description
            types:
               type: "string"
               description: Description
            severity:
               type: "string"
               description: Description